third_party/WebKit/LayoutTests/http/tests/priority-hints/link-dynamic-insertion.html

<script src=resources/common.js></script>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<script>
  const tests = [
    {test: async_test('high importance on <link rel=stylesheet>s not fetched by the preload scanner must have no effect on resource load priority'), importance: 'high', expected_priority: kVeryHigh},
    {test: async_test('low importance on <link rel=stylesheet>s not fetched by the preload scanner must have no effect on resource load priority'), importance: 'low', expected_priority: kVeryHigh},
    {test: async_test('auto importance on <link rel=stylesheet>s not fetched by the preload scanner must have no effect on resource load priority'), importance: 'auto', expected_priority: kVeryHigh},
    {test: async_test('invalid importance on <link rel=stylesheet>s not fetched by the preload scanner must have no effect on resource load priority'), importance: 'xyz', expected_priority: kVeryHigh},
    {test: async_test('missing importance on <link rel=stylesheet>s not fetched by the preload scanner must have no effect on resource load priority'), expected_priority: kVeryHigh}
  ];

  let iteration = 0;
  for (const test of tests) {
    const link = document.createElement('link');
    link.rel = 'stylesheet';
    if (test.importance) link.importance = test.importance;
    link.href = `../resources/dummy.css?${iteration++}`;
    link.onload = assert_priority_onload(link.href, test.expected_priority, test.test);
    document.head.appendChild(link);
  }
</script>
